+++ src/components/Cover.jsx
@@ -6,37 +6,27 @@ import styled from 'styled-components';
 
 const Cover = (props) => {
   const Div = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
+    overflow: hidden;
+    display: inline-block;
+    position: relative;
+    min-width: 100%;
+    min-height: 100%;
+    margin: 2em 0;
     padding: 1em;
-    background: url(${(props) => props.backgrounds.default}) 40% 40% / cover no-repeat;
-
-    ${(props) => {
-      let toRender = '';
-
-      for (let key in props.backgrounds) {
-        const img = props.backgrounds[key];
-        toRender += `
-          @media screen and ${props.theme.devices[key]} {
-            background-image: url(${img});
-          }
-        `;
-      }
-
-      return toRender;
-    }}
   `;
 
   const Inner = styled.div`
     margin: 2em ${(props) => props.theme.layout.marginPhone};
+    padding: 0.5em 0 0;
 
     @media screen and ${(props) => props.theme.devices.phoneLarge} {
       margin: 2em ${(props) => props.theme.layout.marginTablet};
+      padding: 2em 0;
     }
 
     @media screen and ${(props) => props.theme.devices.tablet} {
       margin: 6em ${(props) => props.theme.layout.marginDesktop};
+      padding: 2em 0;
     }
   `;
 
@@ -53,11 +43,13 @@ const Cover = (props) => {
   `;
 
   return (
-    <Div backgrounds={{ ...props.backgrounds }}>
+    <Div>
+      {props.imgTop}
       <Inner>
         <Title>{props.title}</Title>
         <Content>{props.children}</Content>
       </Inner>
+      {props.imgBottom}
     </Div>
   );
 };