+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.4.0] - 2021-01-30
+
+### Changed
+
+- Background images to use `gatsby-image`.
+
 ## [0.3.2] - 2021-01-27
 
 ### Fixed
@@ -44,7 +50,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto.github.io/compare/v0.3.2...HEAD
+[unreleased]: https://github.com/visiosto/visiosto.github.io/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/visiosto/visiosto.github.io/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/visiosto/visiosto.github.io/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/visiosto/visiosto.github.io/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/visiosto/visiosto.github.io/compare/v0.2.0...v0.3.0